+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Fixes and cleanups
 
 For the PHPUnit parallel run:
 
+* Only add the Scribunto Sandbox test to the split groups if
+  Scribunto is present.
+  `T368783 <https://phabricator.wikimedia.org/T368783>`_
 * Add a log notice to the end of parallel runs to support developers
   who run into issues with the new config.
   `T361190 <https://phabricator.wikimedia.org/T361190>`_